refactor(types): move CollectionItem to a shared types module
CollectionItem was defined in @/app/collections/page and imported by 11
services / stores / components / types, coupling the shared types layer to a
client page module (and creating a page -> media-item-card -> page import
cycle).

Move the interface to @/types/media-and-posters/collection-item (next to the
MediaItem type it references and mirroring the backend models.CollectionItem
location) and repoint all importers. collections/page.tsx keeps a
backward-compatible re-export so any remaining "@/app/collections/page"
consumers keep working during migration; it can be dropped once they migrate.

No behavior change.
